Исправлены замечания Карины
