+++ app/Http/Controllers/Admin/ActionController.php
@@ -4,13 +4,7 @@ namespace App\Http\Controllers\Admin;
 
 use App\Http\COntrollers\Controller;
 use Illuminate\Http\Request;
-use App\Models\Action;
-use App\Models\City;
-use App\Models\Brand;
-use App\Models\Category;
-use App\Models\Tag;
-use App\Models\ActionCategory;
-use App\Models\ActionTag;
+use App\Models\{Action, City, Brand, Category, Tag};
 
 
 class ActionController extends Controller
@@ -37,20 +31,15 @@ class ActionController extends Controller
     {  
         $image = session('image');
         $image_src = session('image_src');
-        if ($image) {
-            $image_id = $image->id;
-            $image_path = $image->path;
-        }
+       
         return view('adminlte.layouts.primary', [
-            'page' => 'adminlte.parts.action_create',
+            'page' => 'adminlte.pages.action_create',
             'h1' => 'Добавление акции',
             'categories' => Category::all(),
             'brands' => Brand::all(),
             'cities' => City::all(),
             'tags' => Tag::all(),
-            'image_id' => $image_id ?? '',
-            'image_path' => $image_path ?? '',
-            'image_src' => $image_src ?? ''
+            'image' => $image
         ]); 
     }
 
@@ -60,9 +49,6 @@ class ActionController extends Controller
      */
     public function store(Request $request)
     {	
-        $image_src = session('image_src');
-
-        // Сохраняем акцию в базу
 
         $action = Action::create([
         	'name' => $request->name,
@@ -77,28 +63,11 @@ class ActionController extends Controller
         	'type' => $request->type
         ]);
 
-        // Сохраняем в таблицу action_category
+        // Сохраняем в таблицы action_category и action_tag
 
-        $categories_id = $request->categories;
+        $action->categories()->attach($request->categories);
 
-        foreach ($categories_id as $category_id) {
-
-        	ActionCategory::create([
-        		'action_id' => $action->id,
-        		'category_id' => $category_id
-        	]);
-        }
-
-        // Сохраняем в таблицу action_tag
-
-        $tags_id = $request->tags;
-
-        foreach ($tags_id as $tag_id) {
-        	ActionTag::create([
-        		'action_id' => $action->id,
-        		'tag_id' => $tag_id
-        	]);
-        }
+        $action->tags()->attach($request->tags);
 
         return redirect()->route('admin.actions');
     }
@@ -108,27 +77,18 @@ class ActionController extends Controller
     {
         $action = Action::findOrFail($id);
 
-        $actionCategoryCollection = ActionCategory::where('action_id', $id)->get();
+        $actionCategoryArray = $action->categories->pluck('id')->toArray();
 
-        foreach ($actionCategoryCollection as $item) {
-            $actionCategoryArray[] = $item->category_id;
-        }
-
-        $actionTagCollection = ActionTag::where('action_id', $id)->get();
-
-        foreach ($actionTagCollection as $item) {
-            $actionTagArray[] = $item->tag_id;
-        }
+        $actionTagArray = $action->tags->pluck('id')->toArray();
 
         $image = session('image');
-        $image_src = session('image_src');
+      
         if ($image) {
             $image_id = $image->id;
-            $image_path = $image->path;
         }
 
          return view('adminlte.layouts.primary', [
-            'page' => 'adminlte.parts.action_edit',
+            'page' => 'adminlte.pages.action_edit',
             'h1' => 'Редактирование акции',
             'action' => $action,
             'categories' => Category::all(),
@@ -138,8 +98,7 @@ class ActionController extends Controller
             'actionCategoryArray' => $actionCategoryArray ?? [],
             'actionTagArray' => $actionTagArray ?? [],
             'image_id' => $image_id ?? $action->upload->id,
-            'image_path' => $image_path ?? '',
-            'image_src' => $image_src ?? 'this_img'
+            'image' => $image
 
         ]); 
     }
@@ -151,66 +110,17 @@ class ActionController extends Controller
         $image = session('image');
         $image_src = session('image_src');
 
-        $action->name = $request->name;
-        $action->brand_id = $request->brand;
-        $action->city_id = $request->city;
-        $action->active_from = $request->active_from;
-        $action->active_to = $request->active_to;
-        $action->text = $request->text;
-        $action->links = $request->links;
-        $action->upload_id = $request->image_id;
-        $action->status = "approved";
-        $action->type = $request->type;
-        $action->save();
+        $request->merge(['upload_id' => $request->image_id]);
 
-        // Сохраняем в таблицу action_category
-
-        $request_categories_id = $request->categories;
-
-        $old_categories = ActionCategory::where('action_id', $action->id)->get();
-        $old_categories_id = [];
-
-        $categories_id_to_add = [];
+        $action->update($request->all()); 
        
-        foreach ($old_categories as $item) {
+        // Сохраняем в таблицу action_category и action_tag
 
-            $old_categories_id[] = $item->category_id;
+        $action->categories()->sync($request->categories);
+       
+        $action->tags()->sync($request->tags);
 
-            if (in_array($item->category_id, $request_categories_id)) {  // если старая категория есть в реквесте
-                
-                $categories_id_to_add[] = $item->category_id; 
-          
-            } else {                                                    // если старую категорию при редактировании убрали
-                $category_to_del = ActionCategory::where(['action_id' => $action->id, 'category_id' => $item->category_id])->delete();
-            }
-        }
-
-        foreach ($request_categories_id as $item) {
-            if (!in_array($item, $old_categories_id)) {
-                $categories_id_to_add[] = $item;
-            }
-        }
-        
-        foreach ($categories_id_to_add as $category_id) {
-            $action_category = ActionCategory::firstOrCreate([
-                'action_id' => $action->id,
-                'category_id' => $category_id
-            ]);
-        }
-
-         // Сохраняем в таблицу action_tag
-
-         $tag_to_del = ActionTag::where('action_id', $action->id)->delete();
-
-         $request_tags_id = $request->tags;
-
-         foreach ($request_tags_id as $item) {
-            $action_tag = ActionTag::create([
-                'action_id' => $action->id,
-                'tag_id' => $item
-            ]);
-         }
-    
+         
         return redirect()->route('admin.actions');
     }
