+++ app/helpers.php
@@ -1,0 +1,6 @@
+<?php
+
+function dateFormat($time = null) {
+
+	return \Carbon\Carbon::parse($time)->format('Y-m-d');
+}
