+++ public/css/extra-style.css
@@ -1,0 +1,37 @@
+.img-wrapper {
+	width: 300px;
+	height: 300px;
+	overflow: hidden;
+	margin-bottom: 10px;
+}
+
+.img-wrapper-min {
+	width: 150px;
+	height: 100px;
+	overflow: hidden;
+}
+
+.img-wrapper img, .img-wrapper-min img  {
+	width: 100%;
+}
+
+#upload-demo {
+	width: 350px;
+}
+
+#upload-demo-i {
+	background: #e1e1e1;
+	width: 300px;
+	padding: 30px;
+	height: 300px;
+	margin-top: 30px
+}
+
+.upload-input {
+	padding-top: 30px;
+} 
+
+button.upload-result {
+	float: right;
+	margin-top: 205px;
+}
